Add Constructor Method for reusing token
I am storing the Token information. This allows me to create a VaultApi object without having to get a new AccessToken. I would appreciate it if you could add this to your code. This way I will not have to manually add it back in when downloading your latest changes.

Thank You.