Bug Fix - Retrieving Result Causes a Deadlock
I was able to retrieve the Token when running in a console application, but I had to make the change above to get it to work with a web application. This is due to web applications and Web APIs using contexts. A console application doesn't use contexts; hence, why it worked in unit testing. The following two links go into more detail about this issue...

http://stackoverflow.com/questions/34078296/httpclient-postasync-never-return-response

http://blog.stephencleary.com/2012/07/dont-block-on-async-code.html